Refactored libr/util/date.c
