+++ libr/bin/format/pe/pe.c
@@ -514,66 +514,6 @@ error:
 	return false;
 }
 
-static char *_time_stamp_to_str(ut32 timeStamp) {
-#ifdef _MSC_VER
-	time_t rawtime;
-	struct tm *tminfo;
-	rawtime = (time_t)timeStamp;
-	tminfo = localtime (&rawtime);
-	//tminfo = gmtime (&rawtime);
-	return r_str_trim (strdup (asctime (tminfo)));
-#else
-	struct my_timezone {
-		int tz_minuteswest;     /* minutes west of Greenwich */
-		int tz_dsttime;         /* type of DST correction */
-	} tz;
-	struct timeval tv;
-	int gmtoff;
-	time_t ts = (time_t) timeStamp;
-	gettimeofday (&tv, (void*) &tz);
-	gmtoff = (int) (tz.tz_minuteswest * 60); // in seconds
-	ts += (time_t)gmtoff;
-	return r_str_trim (strdup (ctime (&ts)));
-#endif
-}
-
-static char*_time_stamp_to_dos_str(ut32 timeStamp) {
-        ut16 date = timeStamp >> 16; 
-        ut16 time = timeStamp & 0xFFFF;
-
-        /* Date */
-        ut32 year = ((date & 0xfe00) >> 9) + 1980;
-        ut32 month = (date & 0x01e0) >> 5;
-        ut32 day = date & 0x001f;
-
-        /* Time */
-        ut32 hour = (time & 0xf800) >> 11; 
-        ut32 minutes = (time & 0x07e0) >> 5;
-        ut32 seconds = (time & 0x001f) << 1;
-
-	/* Convert to epoch */
-	struct tm t = {0};
-	t.tm_year = year - 1900;
-	t.tm_mon = month - 1;
-	t.tm_mday = day;
-	t.tm_hour = hour;
-	t.tm_min = minutes;
-	t.tm_sec = seconds;
-	t.tm_isdst = -1;
-	time_t epochTime = mktime (&t);
-
-	/* We reuse the exising function to have consistent behaviour with the other timestamp */
-	return _time_stamp_to_str ((ut32) epochTime);
-}
-
-static bool _time_stamp_is_dos_format(const ut32 compileTimeStamp, const ut32 resourceTimeStamp) {
-	/* We assume they're both POSIX timestamp and thus the higher bits would be equal if they're close to each other */
-	if ((compileTimeStamp >> 16) == (resourceTimeStamp >> 16)) {
-		return false;
-	}
-	return true;
-}
-
 static int bin_pe_init_hdr(struct PE_(r_bin_pe_obj_t)* bin) {
 	if (!(bin->dos_header = malloc (sizeof(PE_(image_dos_header))))) {
 		r_sys_perror ("malloc (dos header)");
@@ -650,7 +590,7 @@ static int bin_pe_init_hdr(struct PE_(r_bin_pe_obj_t)* bin) {
 	// adding compile time to the SDB
 	{
 		sdb_num_set (bin->kv, "image_file_header.TimeDateStamp", bin->nt_headers->file_header.TimeDateStamp, 0);
-		char *timestr = _time_stamp_to_str (bin->nt_headers->file_header.TimeDateStamp);
+		char *timestr = r_time_stamp_to_str (bin->nt_headers->file_header.TimeDateStamp);
 		sdb_set_owned (bin->kv, "image_file_header.TimeDateStamp_string", timestr, 0);
 	}
 	bin->optional_header = &bin->nt_headers->optional_header;
@@ -2398,10 +2338,10 @@ static void _parse_resource_directory(struct PE_(r_bin_pe_obj_t) *bin, Pe_image_
 			break;
 		}
 		/* Compare compileTimeStamp to resource timestamp to figure out if DOS date or POSIX date */
-		if (_time_stamp_is_dos_format ((ut32) sdb_num_get (bin->kv, "image_file_header.TimeDateStamp", 0), dir->TimeDateStamp)) {
-			rs->timestr = _time_stamp_to_dos_str (dir->TimeDateStamp);
+		if (r_time_stamp_is_dos_format ((ut32) sdb_num_get (bin->kv, "image_file_header.TimeDateStamp", 0), dir->TimeDateStamp)) {
+			rs->timestr = r_time_stamp_to_str ( r_dos_time_stamp_to_posix (dir->TimeDateStamp));
 		} else {
-			rs->timestr = _time_stamp_to_str (dir->TimeDateStamp);
+			rs->timestr = r_time_stamp_to_str (dir->TimeDateStamp);
 		}
 		rs->type = _resource_type_str (type);
 		rs->language = strdup (_resource_lang_str (entry.u1.Name & 0x3ff));
