+++ libr/include/r_util.h
@@ -35,6 +35,7 @@ int gettimeofday (struct timeval* p, void* tz);
 #include "r_util/r_buf.h"
 #include "r_util/r_bitmap.h"
 #include "r_util/r_constr.h"
+#include "r_util/r_date.h"
 #include "r_util/r_debruijn.h"
 #include "r_util/r_cache.h"
 #include "r_util/r_ctypes.h"
